Enable alternative policy model

+++ functions/taxes.jl
@@ -4,8 +4,7 @@ function street_light(energy::Float64, cost::Float64)
 
 end
 
-function VAT(energy::Float64, cost::Float64; rate::Float64 = 0.13, cutoff::Int64 = 280)
-    
+function VAT(energy::Float64, cost::Float64; rate::Float64 = 0.13, cutoff = 280)
     tax = 0.0
     
     if energy > cutoff
